requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.2
python-dotenv==1.2.2
groq==1.2.0